+++ android/src/main/java/com/getcapacitor/community/database/sqlite/SQLite/ImportExportJson/ImportFromJson.java
@@ -13,6 +13,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 import org.json.JSONException;
+import org.json.JSONObject;
 
 public class ImportFromJson {
 
@@ -234,11 +235,11 @@ public class ImportFromJson {
                 .append(" AFTER UPDATE ON ")
                 .append(tableName)
                 .append(" FOR EACH ROW ")
-                .append("WHEN NEW.last_modified < " + "OLD.last_modified BEGIN ")
+                .append("WHEN NEW.last_modified <= " + "OLD.last_modified BEGIN ")
                 .append("UPDATE ")
                 .append(tableName)
                 .append(" SET last_modified = (strftime('%s','now')) ")
-                .append("WHERE id=OLD.id; ")
+                .append("WHERE id=NEW.id; ")
                 .append("END;")
                 .toString();
             statements.add(stmtTrigger);
@@ -380,42 +381,52 @@ public class ImportFromJson {
             if (tableNamesTypes.length() == 0) {
                 throw new Exception("CreateTableData: no column names & types returned");
             }
-            ArrayList<String> tColNames = new ArrayList<>();
-            ArrayList<String> tColTypes = new ArrayList<>();
+            ArrayList<String> tColNames;
             if (tableNamesTypes.has("names")) {
                 tColNames = _uJson.getColumnNames(tableNamesTypes.get("names"));
             } else {
                 throw new Exception("GetValues: Table " + tableName + " no names");
             }
-            if (tableNamesTypes.has("types")) {
-                tColTypes = _uJson.getColumnNames(tableNamesTypes.get("types"));
-            } else {
-                throw new Exception("GetValues: Table " + tableName + " no types");
+            // New process flow
+            JSONObject retObjStrs = generateInsertAndDeletedStrings(tColNames,values);
+            // Create the statement for INSERT
+            String namesString = _uJson.convertToString(tColNames, ',');
+            if(retObjStrs.has("insert")) {
+              String stmtInsert =
+                new StringBuilder("INSERT OR REPLACE INTO ")
+                  .append(tableName)
+                  .append("(")
+                  .append(namesString)
+                  .append(") ")
+                  .append(retObjStrs.get("insert"))
+                  .append(";")
+                  .toString();
+              JSObject retObj = mDb.prepareSQL(stmtInsert, new ArrayList<>(), true,
+                "no");
+              long lastId = retObj.getLong("lastId");
+              if (lastId < 0) {
+                throw new Exception("CreateTableData: INSERT lastId < 0");
+              }
+            }
+            if(retObjStrs.has("delete")) {
+              String stmtDelete =
+                new StringBuilder("DELETE FROM ")
+                  .append(tableName)
+                  .append(" WHERE ")
+                  .append(tColNames.get(0))
+                  .append(" ")
+                  .append(retObjStrs.get("delete"))
+                  .append(";")
+                  .toString();
+              JSObject retObj = mDb.prepareSQL(stmtDelete, new ArrayList<>(), true,
+                "no");
+              long lastId = retObj.getLong("lastId");
+              if (lastId < 0) {
+                throw new Exception("CreateTableData: INSERT lastId < 0");
+              }
             }
 
-            // Loop on Table's Values
-            for (int j = 0; j < values.size(); j++) {
-                // Check the row number of columns
-                ArrayList<Object> row = values.get(j);
-                // Check row validity remove to accept RDBMS types
-                //                _uJson.checkRowValidity(mDb, tColNames, tColTypes, row, j, tableName);
-                // Create INSERT or UPDATE Statements
-                Boolean isRun = true;
-                String stmt = createRowStatement(mDb, tColNames, tColTypes, row, j, tableName, mode);
-                isRun = checkUpdate(mDb, stmt, row, tableName, tColNames, tColTypes);
-                if (isRun) {
-                    // load the values
-                    if (stmt.substring(0, 6).toUpperCase().equals("DELETE")) {
-                        row = new ArrayList<>();
-                    }
-                    JSObject retObj = mDb.prepareSQL(stmt, row, true, "no");
-                    long lastId = retObj.getLong("lastId");
-                    if (lastId < 0) {
-                        throw new Exception("CreateTableData: lastId < 0");
-                    }
-                }
-            }
-            return;
+
         } catch (JSONException e) {
             throw new Exception("CreateTableData: " + e.getMessage());
         } catch (Exception e) {
@@ -424,173 +435,72 @@ public class ImportFromJson {
     }
 
     /**
-     * Check when UPDATE if the values are updated
-     * @param mDb
-     * @param stmt
-     * @param values
-     * @param tableName
+     * GenerateInsertAndDeletedStrings
      * @param tColNames
-     * @param tColTypes
-     * @return
-     * @throws Exception
-     */
-    private Boolean checkUpdate(
-        Database mDb,
-        String stmt,
-        ArrayList<Object> values,
-        String tableName,
-        ArrayList<String> tColNames,
-        ArrayList<String> tColTypes
-    ) throws Exception {
-        Boolean isRun = true;
-        if (stmt.substring(0, 6).equals("UPDATE")) {
-            StringBuilder sbQuery = new StringBuilder("SELECT * FROM ").append(tableName).append(" WHERE ").append(tColNames.get(0));
-
-            if (values.get(0) instanceof String) {
-                sbQuery.append(" = '").append(values.get(0)).append("';");
-            } else {
-                sbQuery.append(" = ").append(values.get(0)).append(";");
-            }
-            String query = sbQuery.toString();
-
-            try {
-                ArrayList<ArrayList<Object>> resValues = _uJson.getValues(mDb, query, tableName);
-                if (resValues.size() > 0) {
-                    isRun = checkValues(values, resValues.get(0));
-                } else {
-                    throw new Exception("CheckUpdate: CheckUpdate statement returns nothing");
-                }
-            } catch (Exception e) {
-                throw new Exception("CheckUpdate: " + e.getMessage());
-            }
-        }
-        return isRun;
-    }
-
-    /**
-     * Check Values
      * @param values
-     * @param nValues
      * @return
-     * @throws Exception
+     * @throws JSONException
      */
-    private Boolean checkValues(ArrayList<Object> values, ArrayList<Object> nValues) throws Exception {
-        if (values.size() > 0 && nValues.size() > 0 && values.size() == nValues.size()) {
-            for (int i = 0; i < values.size(); i++) {
-                if (nValues.get(i) instanceof String) {
-                    if (!values.get(i).equals(nValues.get(i))) {
-                        return true;
-                    }
-                } else if (nValues.get(i) instanceof Long && values.get(i) instanceof Integer) {
-                    //            int iVal = (Integer) values.get(i);
-                    long lVal = (Integer) values.get(i);
-                    //            long nlVal = (Long) nValues.get(i);
-                    if (lVal != (Long) nValues.get(i)) {
-                        return true;
-                    }
-                } else if (nValues.get(i) instanceof Double && values.get(i) instanceof Integer) {
-                    double dVal = (Integer) values.get(i);
-                    if (dVal != (Double) nValues.get(i)) {
-                        return true;
-                    }
-                } else {
-                    if (values.get(i) != nValues.get(i)) {
-                        return true;
-                    }
-                }
+    private JSONObject generateInsertAndDeletedStrings(ArrayList<String> tColNames,
+                                              ArrayList<ArrayList<Object>> values)
+                                                            throws JSONException {
+      JSONObject retObj = new JSONObject();
+      StringBuilder insertValues = new StringBuilder();
+      StringBuilder deletedIds = new StringBuilder();
+
+      for (ArrayList<Object> rowIndex : values) {
+        int colIndex = tColNames.indexOf("sql_deleted");
+
+        // Check if the column "sql_deleted" is 0
+        if (colIndex == -1 || (int) rowIndex.get(colIndex) == 0) {
+
+          if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.N) {
+            String formattedRow = null;
+            formattedRow = String.join(", ", rowIndex.stream().map(item -> {
+              if (item instanceof String) {
+                return "'" + item + "'";
+              } else {
+                return item.toString();
+              }
+            }).toArray(String[]::new));
+            insertValues.append("(").append(formattedRow).append("), ");
+          } else {
+            StringBuilder formattedRow = new StringBuilder();
+            for (int i = 0; i < rowIndex.size(); i++) {
+              if (i > 0) {
+                formattedRow.append(", ");
+              }
+              Object item = rowIndex.get(i);
+              if (item instanceof String) {
+                formattedRow.append("'").append(item).append("'");
+              } else {
+                formattedRow.append(item);
+              }
             }
-            return false;
-        } else {
-            throw new Exception("CheckValues: Both arrays not the same length");
+            insertValues.append("(").append(formattedRow).append("), ");
+          }
+        } else if ((int) rowIndex.get(colIndex) == 1) {
+          deletedIds.append(rowIndex.get(0)).append(", ");
         }
-    }
+      }
 
-    /**
-     * Create the Row Statement to load the data
-     * @param mDb
-     * @param tColNames
-     * @param tColTypes
-     * @param row
-     * @param j
-     * @param tableName
-     * @param mode
-     * @return
-     */
-    private String createRowStatement(
-        Database mDb,
-        ArrayList<String> tColNames,
-        ArrayList<String> tColTypes,
-        ArrayList<Object> row,
-        int j,
-        String tableName,
-        String mode
-    ) throws Exception {
-        String msg = "CreateRowStatement: ";
-        msg += "Table" + tableName + " values row";
-        if (tColNames.size() != row.size() || row.size() == 0 || tColNames.size() == 0) {
-            throw new Exception(msg + j + " not correct length");
-        }
-
-        boolean retIsIdExists = _uJson.isIdExists(mDb, tableName, tColNames.get(0), row.get(0));
-        String stmt = "";
-        // Create INSERT or UPDATE Statements
-        if (mode.equals("full") || (mode.equals("partial") && !retIsIdExists)) {
-            // Insert
-            String namesString = _uJson.convertToString(tColNames, ',');
-            String questionMarkString = _uJson.createQuestionMarkString(tColNames.size());
-            if (questionMarkString.length() == 0) {
-                throw new Exception(msg + j + "questionMarkString is empty");
-            }
-            stmt =
-                new StringBuilder("INSERT INTO ")
-                    .append(tableName)
-                    .append("(")
-                    .append(namesString)
-                    .append(")")
-                    .append(" VALUES (")
-                    .append(questionMarkString)
-                    .append(");")
-                    .toString();
-        } else {
-            Boolean isUpdate = true;
-            Integer idxDelete = tColNames.indexOf("sql_deleted");
-            if (idxDelete >= 0) {
-                if (row.get(idxDelete).equals(1)) {
-                    // Delete
-                    isUpdate = false;
-                    Object key = tColNames.get(0);
-                    StringBuilder sbQuery = new StringBuilder("DELETE FROM ")
-                        .append(tableName)
-                        .append(" WHERE ")
-                        .append(tColNames.get(0))
-                        .append(" = ");
-
-                    if (key instanceof Integer) sbQuery.append(row.get(0)).append(";");
-                    if (key instanceof String) sbQuery.append("'").append(row.get(0)).append("';");
-                    stmt = sbQuery.toString();
-                }
-            }
-            if (isUpdate) {
-                // Update
-                String setString = _uJson.setNameForUpdate(tColNames);
-                if (setString.length() == 0) {
-                    throw new Exception(msg + j + "setString is empty");
-                }
-                Object key = tColNames.get(0);
-                StringBuilder sbQuery = new StringBuilder("UPDATE ")
-                    .append(tableName)
-                    .append(" SET ")
-                    .append(setString)
-                    .append(" WHERE ")
-                    .append(tColNames.get(0))
-                    .append(" = ");
-
-                if (key instanceof Integer) sbQuery.append(row.get(0)).append(";");
-                if (key instanceof String) sbQuery.append("'").append(row.get(0)).append("';");
-                stmt = sbQuery.toString();
-            }
-        }
-        return stmt;
+      // Remove the trailing comma and space from insertValues and deletedIds
+      if (insertValues.length() > 0) {
+        insertValues.setLength(insertValues.length() - 2); // Remove trailing comma and space
+        insertValues.insert(0, "VALUES ");
+      }
+      if (deletedIds.length() > 0) {
+        deletedIds.setLength(deletedIds.length() - 2); // Remove trailing comma and space
+        deletedIds.insert(0, "IN (");
+        deletedIds.append(")");
+      }
+      if (insertValues.length() > 0) {
+        retObj.put("insert", insertValues.toString());
+      }
+      if (deletedIds.length() > 0) {
+        retObj.put("delete", deletedIds.toString());
+      }
+      return retObj;
     }
 
     /**